% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTemplate.R
\name{addTemplate}
\alias{addTemplate}
\title{Add 'inst/template' directory}
\usage{
addTemplate(base = ".")
}
\arguments{
\item{base}{directory of package (defaults to '.')}
}
\value{
invisible(NULL)
}
\description{
Called for side-effect only of creating 'inst/templates' directory where it
does not already exist.
}
\author{
Michael Nelson
}

