% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTemplate.R
\name{createTemplate}
\alias{createTemplate}
\title{Create template folder structure}
\usage{
createTemplate(name = "newTemplate", description = "A new template",
  yamlextra = list(), titleblockextra = list())
}
\arguments{
\item{name}{name of the template}

\item{description}{a short description of the template}

\item{yamlextra}{list of extra details for template.yaml}

\item{titleblockextra}{list of extra options for title block in rmd}
}
\description{
Creates template folder structure according to http://rmarkdown.rstudio.com/developer_document_templates.html.
This assumes you are working to create an R package
}
\author{
Michael Nelson
}

