% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTemplate.R
\name{makeTitleBlock}
\alias{makeTitleBlock}
\title{A simple function to create a title block for an rmarkdown template.}
\usage{
makeTitleBlock(title = "Untitled", ...)
}
\arguments{
\item{title}{title defaults to 'Untitled'}
}
\description{
A simple function to create a title block for an rmarkdown template.
}

