% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTemplate.R
\name{makeTitleBlock}
\alias{makeTitleBlock}
\title{Create a title block for an rmarkdown template.}
\usage{
makeTitleBlock(title = "Untitled", ...)
}
\arguments{
\item{title}{title defaults to 'Untitled'}

\item{...}{a list of additional options such as output}
}
\description{
Create a title block for an rmarkdown template.
}

