% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTemplate.R
\name{makeTitleBlock}
\alias{makeTitleBlock}
\title{Create a title block for an rmarkdown template.}
\usage{
makeTitleBlock(title = "Untitled", ...)
}
\arguments{
\item{title}{title defaults to 'Untitled'}

\item{...}{a list of additional options such as output}
}
\value{
title block as a character string length 1
}
\description{
Create a title block for an rmarkdown template.
}
\examples{
\dontrun{
# create a title block for an HTML document
makeTitleBlock(title = "Something interesting", output = "html_document")
# create a title block with some extra options
makeTitleBlock(title = "Something more interesting",
  output = list(html_document = list(toc = TRUE, fig_caption = TRUE, css = 'styles.css')))
}
}
\author{
Michael Nelson
}

